favicon / wp 3.0 tweaks
git-svn-id: file:///Users/michaelkoepke/svn/semiologic/plugins/favicon-head/trunk@4414 6c9c31c4-cb3d-4424-af3d-876279169c7e

+++ favicon-head.php
@@ -4,7 +4,7 @@ Plugin Name: Favicon Head
 Plugin URI: http://timjoh.com/wordpress-plugin-favicon-head/
 Description: Favicon Head adds meta tags in the head of every page, specifying the location of your blog's favicon.ico.
 Author: Tim A. Johansson, Denis de Bernardy
-Version: 1.4.1 fork
+Version: 1.5 RC fork
 Author URI: http://timjoh.com/
 
 Copyright 2006  Tim A. Johansson
@@ -44,35 +44,35 @@ define('taj_fha_location_default', '');
 
 add_action('admin_menu', 'taj_fh_add_options_pages');
 
-add_option('taj_fh_location', taj_fh_location_default, 'Default location for the favicon.');
-add_option('taj_fha_location', taj_fha_location_default, 'Default location for the animated favicon.');
+if ( !get_site_option('taj_fh_location') ) {
+	update_site_option('taj_fh_location', get_option('taj_fh_location') ? get_option('taj_fh_location') : '/favicon.ico');
+	update_site_option('taj_fha_location', get_option('taj_fha_location') ? get_option('taj_fha_location') : '');
+}
 
 function taj_fh_add_options_pages() {
-	add_options_page('Favicon', 'Favicon', 'manage_options', __FILE__, 'taj_fh_options_page');
+	if ( function_exists('is_super_admin') && !is_super_admin() )
+		return;
+	add_options_page('Favicon', 'Favicon', 'manage_options', 'favicon-head', 'taj_fh_options_page');
 }
 
 function taj_fh_options_page() {
+	if ( function_exists('is_super_admin') && !is_super_admin() )
+		return;
  	if (isset($_POST['info_update'])) {
 		check_admin_referer('favicon');
 
 		// Update location
-		$taj_fh_location = $_POST['taj_fh_location'];
-		$taj_fha_location = $_POST['taj_fha_location'];
-		if ( ! isset( $taj_fh_location ) ) {
-			$taj_fh_location = taj_fh_location_default;
-		}
-		update_option('taj_fh_location', $taj_fh_location);
-		if ( ! isset( $taj_fha_location ) ) {
-			$taj_fha_location = taj_fha_location_default;
-		}
-		update_option('taj_fha_location', $taj_fha_location);
+		$taj_fh_location = stripslashes($_POST['taj_fh_location']);
+		$taj_fha_location = stripslashes($_POST['taj_fha_location']);
+		update_site_option('taj_fh_location', $taj_fh_location);
+		update_site_option('taj_fha_location', $taj_fha_location);
 
 		// Acknowledge
 		echo '<div class="updated"><p><strong>Favicon Head options updated</strong></p></div>';
 	}
 	?>
 		<div class="wrap">
-			<form method="post" action="options-general.php?page=favicon-head.php">
+			<form method="post" action="options-general.php?page=favicon-head">
 			<?php if ( function_exists('wp_nonce_field') ) wp_nonce_field('favicon'); ?>
 			<h2>Favicon Settings</h2>
 				<table class="form-table">
@@ -81,7 +81,7 @@ function taj_fh_options_page() {
 							<label for="taj_fh_location">Static Favicon</label>
 						</th>
 						<td>
-							<input type="text" name="taj_fh_location" size="60" class="code" value="<?php echo attribute_escape(get_option('taj_fh_location')); ?>" />
+							<input type="text" name="taj_fh_location" size="60" class="code" value="<?php echo esc_attr(get_site_option('taj_fh_location')); ?>" />
 							<p>"/favicon.ico", means that the favicon.ico file is located in the root folder of the site.</p>
 							<p>To override this, enter the url of your favicon file.</p>
 						</td>
@@ -91,7 +91,7 @@ function taj_fh_options_page() {
 							<label for="taj_fha_location">Animated Favicon</label>
 						</th>
 						<td>
-							<input type="text" name="taj_fha_location" size="60" class="code" value="<?php echo attribute_escape(get_option('taj_fha_location')); ?>" />
+							<input type="text" name="taj_fha_location" size="60" class="code" value="<?php echo esc_attr(get_site_option('taj_fha_location')); ?>" />
 							<p>To use an animated favicon, enter its url. Be sure to keep a static favicon around if you do, however: animated favicons do not work in most versions of IE.</p>
 						</td>
 					</tr>
@@ -106,8 +106,8 @@ function taj_fh_options_page() {
 
 function taj_fh_meta() {
 	/* The guidelines in the Wikipedia article are followed. */
-	$favicon_location = get_option('taj_fh_location');
-	$afavicon_location = get_option('taj_fha_location');
+	$favicon_location = esc_attr(get_site_option('taj_fh_location'));
+	$afavicon_location = esc_attr(get_site_option('taj_fha_location'));
 	echo '<link rel="shortcut icon" href="' . $favicon_location . '" type="image/x-icon" />' . "\n"; /* For IE */
 	echo '<link rel="icon" href="' . ( empty( $afavicon_location ) ? $favicon_location : $afavicon_location ) . '" type="image/x-icon" />' . "\n"; /* For sane browsers */
 }
